fix: restore stock behavior
Restores original app behavior lost/altered during jadx reconstruction:
- StateMachine inner classes (StatePhotoReady/StateVideoReady/
  StateVideoStopping/StateVideoStore/StateCaptureCountdown): removed a
  redundant explicit StateMachine constructor parameter that duplicated
  the compiler's implicit outer-class reference.
- BaseRecorderController (Prepare/Start/Stop/Pause/ResumeTask): moved
  the actual work (prepareInternal/startInternal/stopInternal/
  pauseInternal/resumeInternal, notifyError) outside the mStateLock
  synchronized block, matching the original's narrower lock scope.
- SuperSlowRecorderController.startBypassCameraSuperSlow(): same
  synchronized-scope fix for the bypass-camera start path.

+++ app/src/main/java/com/sonyericsson/android/camera/recorder/utility/FpsMonitor.java
@@ -37,7 +37,7 @@ public class FpsMonitor {
     }
 
     private void addResult(long j) {
-        this.mResult.add(Double.valueOf(((double) (this.mSampleCount - 1)) / ((j - this.mHeadSampleTime) / 1.0E9d)));
+        this.mResult.add(Double.valueOf((this.mSampleCount - 1) / ((j - this.mHeadSampleTime) / 1.0E9d)));
     }
 
     public String dump() {
